Practically rewrite dnshook
Make it more readable.

Introduce specific methods for getting reverse dns info of an IPv4
or eui64.

Parse every nic info and check for DNS_FLAG network flag

Do not explicitly pass zone statement to nsupdate. Let it determine
the correct zone to update based on the rest of the input.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>