Add configure args for Ganeti and nfdhcpd paths
Add --with-ganeti-nic-dir and --with-nfdhcpd-state-dir options in
configure. The program needs to access the directory where the nfdhcpd
binding files are stored and the one that Ganeti uses to store the NIC
info for the instances when working with XEN. The parameters that store
this info should be configurable at compile time.